requirements.txt
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.1.0
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
